requirements.txt
flask
pyTelegramBotAPI
python-dotenv
aliexpress-api
requests

